migrate service_api dataset controller tests to testcontainers integration tests
