--- api/tests/unit_tests/controllers/service_api/dataset/test_dataset.py
@@ -1,1522 +1,0 @@
-"""
-Unit tests for Service API Dataset controllers.
-
-Tests coverage for:
-- DatasetCreatePayload, DatasetUpdatePayload Pydantic models
-- Tag-related payloads (create, update, delete, binding)
-- DatasetListQuery model
-- DatasetService and TagService interfaces
-- Permission validation patterns
-
-Focus on:
-- Pydantic model validation
-- Error type mappings
-- Service method interfaces
-"""
-
-import uuid
-from types import SimpleNamespace
-from unittest.mock import Mock, patch
-
-import pytest
-from werkzeug.exceptions import Forbidden, NotFound
-
-import services
-from controllers.service_api.dataset.dataset import (
-    DatasetCreatePayload,
-    DatasetListQuery,
-    DatasetUpdatePayload,
-    TagBindingPayload,
-    TagCreatePayload,
-    TagDeletePayload,
-    TagUnbindingPayload,
-    TagUpdatePayload,
-)
-from controllers.service_api.dataset.error import DatasetInUseError, DatasetNameDuplicateError, InvalidActionError
-from models.account import Account
-from models.dataset import DatasetPermissionEnum
-from models.enums import TagType
-from services.dataset_service import DatasetPermissionService, DatasetService, DocumentService
-from services.tag_service import TagService
-
-
-class TestDatasetCreatePayload:
-    """Test suite for DatasetCreatePayload Pydantic model."""
-
-    def test_payload_with_required_name(self):
-        """Test payload with required name field."""
-        payload = DatasetCreatePayload(name="Test Dataset")
-        assert payload.name == "Test Dataset"
-        assert payload.description == ""
-        assert payload.permission == DatasetPermissionEnum.ONLY_ME
-
-    def test_payload_with_all_fields(self):
-        """Test payload with all fields populated."""
-        payload = DatasetCreatePayload(
-            name="Full Dataset",
-            description="A comprehensive dataset description",
-            indexing_technique="high_quality",
-            permission=DatasetPermissionEnum.ALL_TEAM,
-            provider="vendor",
-            embedding_model="text-embedding-ada-002",
-            embedding_model_provider="openai",
-        )
-        assert payload.name == "Full Dataset"
-        assert payload.description == "A comprehensive dataset description"
-        assert payload.indexing_technique == "high_quality"
-        assert payload.permission == DatasetPermissionEnum.ALL_TEAM
-        assert payload.provider == "vendor"
-        assert payload.embedding_model == "text-embedding-ada-002"
-        assert payload.embedding_model_provider == "openai"
-
-    def test_payload_name_length_validation_min(self):
-        """Test name minimum length validation."""
-        with pytest.raises(ValueError):
-            DatasetCreatePayload(name="")
-
-    def test_payload_name_length_validation_max(self):
-        """Test name maximum length validation (40 chars)."""
-        with pytest.raises(ValueError):
-            DatasetCreatePayload(name="A" * 41)
-
-    def test_payload_description_max_length(self):
-        """Test description maximum length (400 chars)."""
-        with pytest.raises(ValueError):
-            DatasetCreatePayload(name="Dataset", description="A" * 401)
-
-    @pytest.mark.parametrize("technique", ["high_quality", "economy"])
-    def test_payload_valid_indexing_techniques(self, technique):
-        """Test valid indexing technique values."""
-        payload = DatasetCreatePayload(name="Dataset", indexing_technique=technique)
-        assert payload.indexing_technique == technique
-
-    def test_payload_with_external_knowledge_settings(self):
-        """Test payload with external knowledge configuration."""
-        payload = DatasetCreatePayload(
-            name="External Dataset", external_knowledge_api_id="api_123", external_knowledge_id="knowledge_456"
-        )
-        assert payload.external_knowledge_api_id == "api_123"
-        assert payload.external_knowledge_id == "knowledge_456"
-
-
-class TestDatasetUpdatePayload:
-    """Test suite for DatasetUpdatePayload Pydantic model."""
-
-    def test_payload_all_optional(self):
-        """Test payload with all fields optional."""
-        payload = DatasetUpdatePayload()
-        assert payload.name is None
-        assert payload.description is None
-        assert payload.permission is None
-
-    def test_payload_with_partial_update(self):
-        """Test payload with partial update fields."""
-        payload = DatasetUpdatePayload(name="Updated Name", description="Updated description")
-        assert payload.name == "Updated Name"
-        assert payload.description == "Updated description"
-
-    def test_payload_with_permission_change(self):
-        """Test payload with permission update."""
-        payload = DatasetUpdatePayload(
-            permission=DatasetPermissionEnum.PARTIAL_TEAM,
-            partial_member_list=[{"user_id": "user_123", "role": "editor"}],
-        )
-        assert payload.permission == DatasetPermissionEnum.PARTIAL_TEAM
-        assert len(payload.partial_member_list) == 1
-
-    def test_payload_name_length_validation(self):
-        """Test name length constraints."""
-        # Minimum is 1
-        with pytest.raises(ValueError):
-            DatasetUpdatePayload(name="")
-
-        # Maximum is 40
-        with pytest.raises(ValueError):
-            DatasetUpdatePayload(name="A" * 41)
-
-
-class TestDatasetListQuery:
-    """Test suite for DatasetListQuery Pydantic model."""
-
-    def test_query_with_defaults(self):
-        """Test query with default values."""
-        query = DatasetListQuery()
-        assert query.page == 1
-        assert query.limit == 20
-        assert query.keyword is None
-        assert query.include_all is False
-        assert query.tag_ids == []
-
-    def test_query_with_all_filters(self):
-        """Test query with all filter fields."""
-        query = DatasetListQuery(
-            page=3, limit=50, keyword="machine learning", include_all=True, tag_ids=["tag1", "tag2", "tag3"]
-        )
-        assert query.page == 3
-        assert query.limit == 50
-        assert query.keyword == "machine learning"
-        assert query.include_all is True
-        assert len(query.tag_ids) == 3
-
-    def test_query_with_tag_filter(self):
-        """Test query with tag IDs filter."""
-        query = DatasetListQuery(tag_ids=["tag_abc", "tag_def"])
-        assert query.tag_ids == ["tag_abc", "tag_def"]
-
-
-class TestTagCreatePayload:
-    """Test suite for TagCreatePayload Pydantic model."""
-
-    def test_payload_with_name(self):
-        """Test payload with required name."""
-        payload = TagCreatePayload(name="New Tag")
-        assert payload.name == "New Tag"
-
-    def test_payload_name_length_min(self):
-        """Test name minimum length (1)."""
-        with pytest.raises(ValueError):
-            TagCreatePayload(name="")
-
-    def test_payload_name_length_max(self):
-        """Test name maximum length (50)."""
-        with pytest.raises(ValueError):
-            TagCreatePayload(name="A" * 51)
-
-    def test_payload_with_unicode_name(self):
-        """Test payload with unicode characters."""
-        payload = TagCreatePayload(name="标签 🏷️ Тег")
-        assert payload.name == "标签 🏷️ Тег"
-
-
-class TestTagUpdatePayload:
-    """Test suite for TagUpdatePayload Pydantic model."""
-
-    def test_payload_with_name_and_id(self):
-        """Test payload with name and tag_id."""
-        payload = TagUpdatePayload(name="Updated Tag", tag_id="tag_123")
-        assert payload.name == "Updated Tag"
-        assert payload.tag_id == "tag_123"
-
-    def test_payload_requires_tag_id(self):
-        """Test that tag_id is required."""
-        with pytest.raises(ValueError):
-            TagUpdatePayload(name="Updated Tag")
-
-
-class TestTagDeletePayload:
-    """Test suite for TagDeletePayload Pydantic model."""
-
-    def test_payload_with_tag_id(self):
-        """Test payload with tag_id."""
-        payload = TagDeletePayload(tag_id="tag_to_delete")
-        assert payload.tag_id == "tag_to_delete"
-
-    def test_payload_requires_tag_id(self):
-        """Test that tag_id is required."""
-        with pytest.raises(ValueError):
-            TagDeletePayload()
-
-
-class TestTagBindingPayload:
-    """Test suite for TagBindingPayload Pydantic model."""
-
-    def test_payload_with_valid_data(self):
-        """Test payload with valid binding data."""
-        payload = TagBindingPayload(tag_ids=["tag1", "tag2"], target_id="dataset_123")
-        assert len(payload.tag_ids) == 2
-        assert payload.target_id == "dataset_123"
-
-    def test_payload_rejects_empty_tag_ids(self):
-        """Test that empty tag_ids are rejected."""
-        with pytest.raises(ValueError) as exc_info:
-            TagBindingPayload(tag_ids=[], target_id="dataset_123")
-        assert "Tag IDs is required" in str(exc_info.value)
-
-    def test_payload_single_tag_id(self):
-        """Test payload with single tag ID."""
-        payload = TagBindingPayload(tag_ids=["single_tag"], target_id="dataset_456")
-        assert payload.tag_ids == ["single_tag"]
-
-
-class TestTagUnbindingPayload:
-    """Test suite for TagUnbindingPayload Pydantic model."""
-
-    def test_payload_with_valid_data(self):
-        """Test payload with valid unbinding data."""
-        payload = TagUnbindingPayload(tag_id="tag_123", target_id="dataset_456")
-        assert payload.tag_id == "tag_123"
-        assert payload.target_id == "dataset_456"
-
-
-class TestDatasetTagsApi:
-    """Test suite for DatasetTagsApi endpoints."""
-
-    @pytest.fixture
-    def app(self):
-        """Create Flask test application."""
-        from flask import Flask
-
-        app = Flask(__name__)
-        app.config["TESTING"] = True
-        return app
-
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    def test_get_tags_success(self, mock_tag_service, mock_current_user, app):
-        """Test successful retrieval of dataset tags."""
-        # Arrange - mock_current_user needs to pass isinstance(current_user, Account)
-        from models.account import Account
-
-        mock_account = Mock(spec=Account)
-        mock_account.current_tenant_id = "tenant_123"
-        # Replace the mock with our properly specced one
-        from controllers.service_api.dataset import dataset as dataset_module
-
-        original_current_user = dataset_module.current_user
-        dataset_module.current_user = mock_account
-
-        mock_tag = Mock()
-        mock_tag.id = "tag_1"
-        mock_tag.name = "Test Tag"
-        mock_tag.type = TagType.KNOWLEDGE
-        mock_tag.binding_count = "0"  # Required for Pydantic validation - must be string
-        mock_tag_service.get_tags.return_value = [mock_tag]
-
-        from controllers.service_api.dataset.dataset import DatasetTagsApi
-
-        try:
-            # Act
-            with app.test_request_context("/", method="GET"):
-                api = DatasetTagsApi()
-                response, status_code = api.get("tenant_123")
-
-            # Assert
-            assert status_code == 200
-            assert len(response) == 1
-            assert response[0]["id"] == "tag_1"
-            assert response[0]["name"] == "Test Tag"
-            mock_tag_service.get_tags.assert_called_once_with("knowledge", "tenant_123")
-        finally:
-            dataset_module.current_user = original_current_user
-
-    @pytest.mark.skip(reason="Production code bug: binding_count should be string, not integer")
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    @patch("controllers.service_api.dataset.dataset.service_api_ns")
-    def test_create_tag_success(self, mock_service_api_ns, mock_tag_service, app):
-        """Test successful creation of a dataset tag."""
-        # Arrange
-        from controllers.service_api.dataset import dataset as dataset_module
-        from models.account import Account
-
-        mock_account = Mock(spec=Account)
-        mock_account.has_edit_permission = True
-        mock_account.is_dataset_editor = False
-        original_current_user = dataset_module.current_user
-        dataset_module.current_user = mock_account
-
-        mock_tag = Mock()
-        mock_tag.id = "new_tag_1"
-        mock_tag.name = "New Tag"
-        mock_tag.type = TagType.KNOWLEDGE
-        mock_tag_service.save_tags.return_value = mock_tag
-        mock_service_api_ns.payload = {"name": "New Tag"}
-
-        from controllers.service_api.dataset.dataset import DatasetTagsApi
-
-        try:
-            # Act
-            with app.test_request_context("/", method="POST", json={"name": "New Tag"}):
-                api = DatasetTagsApi()
-                response, status_code = api.post("tenant_123")
-
-            # Assert
-            assert status_code == 200
-            assert response["id"] == "new_tag_1"
-            assert response["name"] == "New Tag"
-            assert response["binding_count"] == 0
-        finally:
-            dataset_module.current_user = original_current_user
-
-    def test_create_tag_forbidden(self, app):
-        """Test tag creation without edit permissions."""
-        # Arrange
-        from werkzeug.exceptions import Forbidden
-
-        from controllers.service_api.dataset import dataset as dataset_module
-        from models.account import Account
-
-        mock_account = Mock(spec=Account)
-        mock_account.has_edit_permission = False
-        mock_account.is_dataset_editor = False
-        original_current_user = dataset_module.current_user
-        dataset_module.current_user = mock_account
-
-        from controllers.service_api.dataset.dataset import DatasetTagsApi
-
-        try:
-            # Act & Assert
-            with app.test_request_context("/", method="POST"):
-                api = DatasetTagsApi()
-                with pytest.raises(Forbidden):
-                    api.post("tenant_123")
-        finally:
-            dataset_module.current_user = original_current_user
-
-    @pytest.mark.skip(reason="Production code bug: binding_count should be string, not integer")
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    @patch("controllers.service_api.dataset.dataset.service_api_ns")
-    def test_update_tag_success(self, mock_service_api_ns, mock_tag_service, app):
-        """Test successful update of a dataset tag."""
-        # Arrange
-        from controllers.service_api.dataset import dataset as dataset_module
-        from models.account import Account
-
-        mock_account = Mock(spec=Account)
-        mock_account.has_edit_permission = True
-        original_current_user = dataset_module.current_user
-        dataset_module.current_user = mock_account
-
-        mock_tag = Mock()
-        mock_tag.id = "tag_1"
-        mock_tag.name = "Updated Tag"
-        mock_tag.type = TagType.KNOWLEDGE
-        mock_tag.binding_count = "5"
-        mock_tag_service.update_tags.return_value = mock_tag
-        mock_tag_service.get_tag_binding_count.return_value = 5
-        mock_service_api_ns.payload = {"name": "Updated Tag", "tag_id": "tag_1"}
-
-        from controllers.service_api.dataset.dataset import DatasetTagsApi
-
-        try:
-            # Act
-            with app.test_request_context("/", method="PATCH", json={"name": "Updated Tag", "tag_id": "tag_1"}):
-                api = DatasetTagsApi()
-                response, status_code = api.patch("tenant_123")
-
-            # Assert
-            assert status_code == 200
-            assert response["id"] == "tag_1"
-            assert response["name"] == "Updated Tag"
-            assert response["binding_count"] == 5
-        finally:
-            dataset_module.current_user = original_current_user
-
-    @pytest.mark.skip(reason="Production code bug: binding_count should be string, not integer")
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    @patch("controllers.service_api.dataset.dataset.service_api_ns")
-    def test_delete_tag_success(self, mock_service_api_ns, mock_tag_service, app):
-        """Test successful deletion of a dataset tag."""
-        # Arrange
-        from controllers.service_api.dataset import dataset as dataset_module
-        from models.account import Account
-
-        mock_account = Mock(spec=Account)
-        mock_account.has_edit_permission = True
-        original_current_user = dataset_module.current_user
-        dataset_module.current_user = mock_account
-
-        mock_tag_service.delete_tag.return_value = None
-        mock_service_api_ns.payload = {"tag_id": "tag_1"}
-
-        from controllers.service_api.dataset.dataset import DatasetTagsApi
-
-        try:
-            # Act
-            with app.test_request_context("/", method="DELETE", json={"tag_id": "tag_1"}):
-                api = DatasetTagsApi()
-                response = api.delete("tenant_123")
-
-            # Assert
-            assert response == ("", 204)
-            mock_tag_service.delete_tag.assert_called_once_with("tag_1")
-        finally:
-            dataset_module.current_user = original_current_user
-
-
-class TestDatasetTagBindingApi:
-    """Test suite for DatasetTagBindingApi endpoints."""
-
-    @pytest.fixture
-    def app(self):
-        """Create Flask test application."""
-        from flask import Flask
-
-        app = Flask(__name__)
-        app.config["TESTING"] = True
-        return app
-
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    @patch("controllers.service_api.dataset.dataset.service_api_ns")
-    def test_bind_tags_success(self, mock_service_api_ns, mock_tag_service, app):
-        """Test successful binding of tags to dataset."""
-        # Arrange
-        from controllers.service_api.dataset import dataset as dataset_module
-        from models.account import Account
-
-        mock_account = Mock(spec=Account)
-        mock_account.has_edit_permission = True
-        mock_account.is_dataset_editor = False
-        original_current_user = dataset_module.current_user
-        dataset_module.current_user = mock_account
-
-        mock_tag_service.save_tag_binding.return_value = None
-        payload = {"tag_ids": ["tag_1", "tag_2"], "target_id": "dataset_123"}
-        mock_service_api_ns.payload = payload
-
-        from controllers.service_api.dataset.dataset import DatasetTagBindingApi
-
-        try:
-            # Act
-            with app.test_request_context("/", method="POST", json=payload):
-                api = DatasetTagBindingApi()
-                response = api.post("tenant_123")
-
-            # Assert
-            assert response == ("", 204)
-            mock_tag_service.save_tag_binding.assert_called_once_with(
-                {"tag_ids": ["tag_1", "tag_2"], "target_id": "dataset_123", "type": "knowledge"}
-            )
-        finally:
-            dataset_module.current_user = original_current_user
-
-    def test_bind_tags_forbidden(self, app):
-        """Test tag binding without edit permissions."""
-        # Arrange
-        from werkzeug.exceptions import Forbidden
-
-        from controllers.service_api.dataset import dataset as dataset_module
-        from models.account import Account
-
-        mock_account = Mock(spec=Account)
-        mock_account.has_edit_permission = False
-        mock_account.is_dataset_editor = False
-        original_current_user = dataset_module.current_user
-        dataset_module.current_user = mock_account
-
-        from controllers.service_api.dataset.dataset import DatasetTagBindingApi
-
-        try:
-            # Act & Assert
-            with app.test_request_context("/", method="POST"):
-                api = DatasetTagBindingApi()
-                with pytest.raises(Forbidden):
-                    api.post("tenant_123")
-        finally:
-            dataset_module.current_user = original_current_user
-
-
-class TestDatasetTagUnbindingApi:
-    """Test suite for DatasetTagUnbindingApi endpoints."""
-
-    @pytest.fixture
-    def app(self):
-        """Create Flask test application."""
-        from flask import Flask
-
-        app = Flask(__name__)
-        app.config["TESTING"] = True
-        return app
-
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    @patch("controllers.service_api.dataset.dataset.service_api_ns")
-    def test_unbind_tag_success(self, mock_service_api_ns, mock_tag_service, app):
-        """Test successful unbinding of tag from dataset."""
-        # Arrange
-        from controllers.service_api.dataset import dataset as dataset_module
-        from models.account import Account
-
-        mock_account = Mock(spec=Account)
-        mock_account.has_edit_permission = True
-        mock_account.is_dataset_editor = False
-        original_current_user = dataset_module.current_user
-        dataset_module.current_user = mock_account
-
-        mock_tag_service.delete_tag_binding.return_value = None
-        payload = {"tag_id": "tag_1", "target_id": "dataset_123"}
-        mock_service_api_ns.payload = payload
-
-        from controllers.service_api.dataset.dataset import DatasetTagUnbindingApi
-
-        try:
-            # Act
-            with app.test_request_context("/", method="POST", json=payload):
-                api = DatasetTagUnbindingApi()
-                response = api.post("tenant_123")
-
-            # Assert
-            assert response == ("", 204)
-            mock_tag_service.delete_tag_binding.assert_called_once_with(
-                {"tag_id": "tag_1", "target_id": "dataset_123", "type": "knowledge"}
-            )
-        finally:
-            dataset_module.current_user = original_current_user
-
-
-class TestDatasetTagsBindingStatusApi:
-    """Test suite for DatasetTagsBindingStatusApi endpoints."""
-
-    @pytest.fixture
-    def app(self):
-        """Create Flask test application."""
-        from flask import Flask
-
-        app = Flask(__name__)
-        app.config["TESTING"] = True
-        return app
-
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    def test_get_dataset_tags_binding_status(self, mock_tag_service, app):
-        """Test retrieval of tags bound to a specific dataset."""
-        # Arrange
-        from controllers.service_api.dataset import dataset as dataset_module
-        from models.account import Account
-
-        mock_account = Mock(spec=Account)
-        mock_account.current_tenant_id = "tenant_123"
-        original_current_user = dataset_module.current_user
-        dataset_module.current_user = mock_account
-
-        mock_tag = Mock()
-        mock_tag.id = "tag_1"
-        mock_tag.name = "Test Tag"
-        mock_tag_service.get_tags_by_target_id.return_value = [mock_tag]
-
-        from controllers.service_api.dataset.dataset import DatasetTagsBindingStatusApi
-
-        try:
-            # Act
-            with app.test_request_context("/", method="GET"):
-                api = DatasetTagsBindingStatusApi()
-                response, status_code = api.get("tenant_123", dataset_id="dataset_123")
-
-            # Assert
-            assert status_code == 200
-            assert response["data"] == [{"id": "tag_1", "name": "Test Tag"}]
-            assert response["total"] == 1
-            mock_tag_service.get_tags_by_target_id.assert_called_once_with("knowledge", "tenant_123", "dataset_123")
-        finally:
-            dataset_module.current_user = original_current_user
-
-
-class TestDocumentStatusApi:
-    """Test suite for DocumentStatusApi batch operations."""
-
-    @pytest.fixture
-    def app(self):
-        """Create Flask test application."""
-        from flask import Flask
-
-        app = Flask(__name__)
-        app.config["TESTING"] = True
-        return app
-
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    @patch("controllers.service_api.dataset.dataset.DocumentService")
-    def test_batch_enable_documents(self, mock_doc_service, mock_dataset_service, app):
-        """Test batch enabling documents."""
-        # Arrange
-        mock_dataset = Mock()
-        mock_dataset_service.get_dataset.return_value = mock_dataset
-        mock_doc_service.batch_update_document_status.return_value = None
-
-        from controllers.service_api.dataset.dataset import DocumentStatusApi
-
-        # Act
-        with app.test_request_context("/", method="PATCH", json={"document_ids": ["doc_1", "doc_2"]}):
-            api = DocumentStatusApi()
-            response, status_code = api.patch("tenant_123", "dataset_123", "enable")
-
-        # Assert
-        assert status_code == 200
-        assert response == {"result": "success"}
-        mock_doc_service.batch_update_document_status.assert_called_once()
-
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_batch_update_dataset_not_found(self, mock_dataset_service, app):
-        """Test batch update when dataset not found."""
-        # Arrange
-        mock_dataset_service.get_dataset.return_value = None
-
-        from werkzeug.exceptions import NotFound
-
-        from controllers.service_api.dataset.dataset import DocumentStatusApi
-
-        # Act & Assert
-        with app.test_request_context("/", method="PATCH", json={"document_ids": ["doc_1"]}):
-            api = DocumentStatusApi()
-            with pytest.raises(NotFound) as exc_info:
-                api.patch("tenant_123", "dataset_123", "enable")
-            assert "Dataset not found" in str(exc_info.value)
-
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    @patch("controllers.service_api.dataset.dataset.DocumentService")
-    def test_batch_update_permission_error(self, mock_doc_service, mock_dataset_service, app):
-        """Test batch update with permission error."""
-        # Arrange
-        mock_dataset = Mock()
-        mock_dataset_service.get_dataset.return_value = mock_dataset
-        from services.errors.account import NoPermissionError
-
-        mock_dataset_service.check_dataset_permission.side_effect = NoPermissionError("No permission")
-
-        from werkzeug.exceptions import Forbidden
-
-        from controllers.service_api.dataset.dataset import DocumentStatusApi
-
-        # Act & Assert
-        with app.test_request_context("/", method="PATCH", json={"document_ids": ["doc_1"]}):
-            api = DocumentStatusApi()
-            with pytest.raises(Forbidden):
-                api.patch("tenant_123", "dataset_123", "enable")
-
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    @patch("controllers.service_api.dataset.dataset.DocumentService")
-    def test_batch_update_invalid_action(self, mock_doc_service, mock_dataset_service, app):
-        """Test batch update with invalid action error."""
-        # Arrange
-        mock_dataset = Mock()
-        mock_dataset_service.get_dataset.return_value = mock_dataset
-        mock_doc_service.batch_update_document_status.side_effect = ValueError("Invalid action")
-
-        from controllers.service_api.dataset.dataset import DocumentStatusApi
-        from controllers.service_api.dataset.error import InvalidActionError
-
-        # Act & Assert
-        with app.test_request_context("/", method="PATCH", json={"document_ids": ["doc_1"]}):
-            api = DocumentStatusApi()
-            with pytest.raises(InvalidActionError):
-                api.patch("tenant_123", "dataset_123", "invalid_action")
-
-    """Test DatasetPermissionEnum values."""
-
-    def test_only_me_permission(self):
-        """Test ONLY_ME permission value."""
-        assert DatasetPermissionEnum.ONLY_ME is not None
-
-    def test_all_team_permission(self):
-        """Test ALL_TEAM permission value."""
-        assert DatasetPermissionEnum.ALL_TEAM is not None
-
-    def test_partial_team_permission(self):
-        """Test PARTIAL_TEAM permission value."""
-        assert DatasetPermissionEnum.PARTIAL_TEAM is not None
-
-
-class TestDatasetErrors:
-    """Test dataset-related error types."""
-
-    def test_dataset_in_use_error_can_be_raised(self):
-        """Test DatasetInUseError can be raised."""
-        error = DatasetInUseError()
-        assert error is not None
-
-    def test_dataset_name_duplicate_error_can_be_raised(self):
-        """Test DatasetNameDuplicateError can be raised."""
-        error = DatasetNameDuplicateError()
-        assert error is not None
-
-    def test_invalid_action_error_can_be_raised(self):
-        """Test InvalidActionError can be raised."""
-        error = InvalidActionError("Invalid action")
-        assert error is not None
-
-
-class TestDatasetService:
-    """Test DatasetService interface methods."""
-
-    def test_get_datasets_method_exists(self):
-        """Test DatasetService.get_datasets exists."""
-        assert hasattr(DatasetService, "get_datasets")
-
-    def test_get_dataset_method_exists(self):
-        """Test DatasetService.get_dataset exists."""
-        assert hasattr(DatasetService, "get_dataset")
-
-    def test_create_empty_dataset_method_exists(self):
-        """Test DatasetService.create_empty_dataset exists."""
-        assert hasattr(DatasetService, "create_empty_dataset")
-
-    def test_update_dataset_method_exists(self):
-        """Test DatasetService.update_dataset exists."""
-        assert hasattr(DatasetService, "update_dataset")
-
-    def test_delete_dataset_method_exists(self):
-        """Test DatasetService.delete_dataset exists."""
-        assert hasattr(DatasetService, "delete_dataset")
-
-    def test_check_dataset_permission_method_exists(self):
-        """Test DatasetService.check_dataset_permission exists."""
-        assert hasattr(DatasetService, "check_dataset_permission")
-
-    def test_check_dataset_model_setting_method_exists(self):
-        """Test DatasetService.check_dataset_model_setting exists."""
-        assert hasattr(DatasetService, "check_dataset_model_setting")
-
-    def test_check_embedding_model_setting_method_exists(self):
-        """Test DatasetService.check_embedding_model_setting exists."""
-        assert hasattr(DatasetService, "check_embedding_model_setting")
-
-    @patch.object(DatasetService, "get_datasets")
-    def test_get_datasets_returns_tuple(self, mock_get):
-        """Test get_datasets returns tuple of datasets and total."""
-        mock_datasets = [Mock(), Mock()]
-        mock_get.return_value = (mock_datasets, 2)
-
-        datasets, total = DatasetService.get_datasets(page=1, per_page=20, tenant_id="tenant_123", user=Mock())
-        assert len(datasets) == 2
-        assert total == 2
-
-    @patch.object(DatasetService, "get_dataset")
-    def test_get_dataset_returns_dataset(self, mock_get):
-        """Test get_dataset returns dataset object."""
-        mock_dataset = Mock()
-        mock_dataset.id = str(uuid.uuid4())
-        mock_dataset.name = "Test Dataset"
-        mock_get.return_value = mock_dataset
-
-        result = DatasetService.get_dataset("dataset_id")
-        assert result.name == "Test Dataset"
-
-    @patch.object(DatasetService, "get_dataset")
-    def test_get_dataset_returns_none_when_not_found(self, mock_get):
-        """Test get_dataset returns None when not found."""
-        mock_get.return_value = None
-
-        result = DatasetService.get_dataset("nonexistent_id")
-        assert result is None
-
-
-class TestDatasetPermissionService:
-    """Test DatasetPermissionService interface."""
-
-    def test_check_permission_method_exists(self):
-        """Test DatasetPermissionService.check_permission exists."""
-        assert hasattr(DatasetPermissionService, "check_permission")
-
-    def test_get_dataset_partial_member_list_method_exists(self):
-        """Test DatasetPermissionService.get_dataset_partial_member_list exists."""
-        assert hasattr(DatasetPermissionService, "get_dataset_partial_member_list")
-
-    def test_update_partial_member_list_method_exists(self):
-        """Test DatasetPermissionService.update_partial_member_list exists."""
-        assert hasattr(DatasetPermissionService, "update_partial_member_list")
-
-    def test_clear_partial_member_list_method_exists(self):
-        """Test DatasetPermissionService.clear_partial_member_list exists."""
-        assert hasattr(DatasetPermissionService, "clear_partial_member_list")
-
-
-class TestDocumentService:
-    """Test DocumentService interface."""
-
-    def test_batch_update_document_status_method_exists(self):
-        """Test DocumentService.batch_update_document_status exists."""
-        assert hasattr(DocumentService, "batch_update_document_status")
-
-
-class TestTagService:
-    """Test TagService interface."""
-
-    def test_get_tags_method_exists(self):
-        """Test TagService.get_tags exists."""
-        assert hasattr(TagService, "get_tags")
-
-    def test_save_tags_method_exists(self):
-        """Test TagService.save_tags exists."""
-        assert hasattr(TagService, "save_tags")
-
-    def test_update_tags_method_exists(self):
-        """Test TagService.update_tags exists."""
-        assert hasattr(TagService, "update_tags")
-
-    def test_delete_tag_method_exists(self):
-        """Test TagService.delete_tag exists."""
-        assert hasattr(TagService, "delete_tag")
-
-    def test_save_tag_binding_method_exists(self):
-        """Test TagService.save_tag_binding exists."""
-        assert hasattr(TagService, "save_tag_binding")
-
-    def test_delete_tag_binding_method_exists(self):
-        """Test TagService.delete_tag_binding exists."""
-        assert hasattr(TagService, "delete_tag_binding")
-
-    def test_get_tags_by_target_id_method_exists(self):
-        """Test TagService.get_tags_by_target_id exists."""
-        assert hasattr(TagService, "get_tags_by_target_id")
-
-    def test_get_tag_binding_count_method_exists(self):
-        """Test TagService.get_tag_binding_count exists."""
-        assert hasattr(TagService, "get_tag_binding_count")
-
-    @patch.object(TagService, "get_tags")
-    def test_get_tags_returns_list(self, mock_get):
-        """Test get_tags returns list of tags."""
-        mock_tags = [
-            Mock(id="tag1", name="Tag One", type="knowledge"),
-            Mock(id="tag2", name="Tag Two", type="knowledge"),
-        ]
-        mock_get.return_value = mock_tags
-
-        result = TagService.get_tags("knowledge", "tenant_123")
-        assert len(result) == 2
-
-    @patch.object(TagService, "save_tags")
-    def test_save_tags_returns_tag(self, mock_save):
-        """Test save_tags returns created tag."""
-        mock_tag = Mock()
-        mock_tag.id = str(uuid.uuid4())
-        mock_tag.name = "New Tag"
-        mock_tag.type = TagType.KNOWLEDGE
-        mock_save.return_value = mock_tag
-
-        result = TagService.save_tags({"name": "New Tag", "type": "knowledge"})
-        assert result.name == "New Tag"
-
-
-class TestDocumentStatusAction:
-    """Test document status action values."""
-
-    def test_enable_action(self):
-        """Test enable action."""
-        action = "enable"
-        assert action in ["enable", "disable", "archive", "un_archive"]
-
-    def test_disable_action(self):
-        """Test disable action."""
-        action = "disable"
-        assert action in ["enable", "disable", "archive", "un_archive"]
-
-    def test_archive_action(self):
-        """Test archive action."""
-        action = "archive"
-        assert action in ["enable", "disable", "archive", "un_archive"]
-
-    def test_un_archive_action(self):
-        """Test un_archive action."""
-        action = "un_archive"
-        assert action in ["enable", "disable", "archive", "un_archive"]
-
-
-# =============================================================================
-# API Endpoint Tests
-#
-# ``DatasetListApi`` and ``DatasetApi`` inherit from ``DatasetApiResource``
-# whose ``method_decorators`` include ``validate_dataset_token``.
-#
-# Decorator strategy:
-# - ``@cloud_edition_billing_rate_limit_check`` preserves ``__wrapped__``
-#   → call via ``_unwrap(method)(self, …)``.
-# - Methods without billing decorators → call directly; only patch ``db``,
-#   services, ``current_user``, and ``marshal``.
-# =============================================================================
-
-
-def _unwrap(method):
-    """Walk ``__wrapped__`` chain to get the original function."""
-    fn = method
-    while hasattr(fn, "__wrapped__"):
-        fn = fn.__wrapped__
-    return fn
-
-
-@pytest.fixture
-def mock_tenant():
-    tenant = Mock()
-    tenant.id = str(uuid.uuid4())
-    return tenant
-
-
-@pytest.fixture
-def mock_dataset():
-    dataset = Mock()
-    dataset.id = str(uuid.uuid4())
-    dataset.tenant_id = str(uuid.uuid4())
-    dataset.indexing_technique = "economy"
-    dataset.embedding_model_provider = None
-    dataset.embedding_model = None
-    return dataset
-
-
-class TestDatasetListApiGet:
-    """Test suite for DatasetListApi.get() endpoint.
-
-    ``get`` has no billing decorators but calls ``current_user``,
-    ``DatasetService``, ``create_plugin_provider_manager``, and ``marshal``.
-    """
-
-    @patch("controllers.service_api.dataset.dataset.marshal")
-    @patch("controllers.service_api.dataset.dataset.create_plugin_provider_manager")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_list_datasets_success(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        mock_provider_mgr,
-        mock_marshal,
-        app,
-        mock_tenant,
-    ):
-        """Test successful dataset list retrieval."""
-        from controllers.service_api.dataset.dataset import DatasetListApi
-
-        mock_current_user.__class__ = Account
-        mock_current_user.current_tenant_id = mock_tenant.id
-        mock_dataset_svc.get_datasets.return_value = ([Mock()], 1)
-
-        mock_configs = Mock()
-        mock_configs.get_models.return_value = []
-        mock_provider_mgr.return_value.get_configurations.return_value = mock_configs
-
-        mock_marshal.return_value = [{"indexing_technique": "economy", "embedding_model_provider": None}]
-
-        with app.test_request_context("/datasets?page=1&limit=20", method="GET"):
-            api = DatasetListApi()
-            response, status = api.get(tenant_id=mock_tenant.id)
-
-        assert status == 200
-        assert "data" in response
-        assert "total" in response
-
-
-class TestDatasetListApiPost:
-    """Test suite for DatasetListApi.post() endpoint.
-
-    ``post`` is wrapped by ``@cloud_edition_billing_rate_limit_check``.
-    """
-
-    @patch("controllers.service_api.dataset.dataset.marshal")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_create_dataset_success(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        mock_marshal,
-        app,
-        mock_tenant,
-    ):
-        """Test successful dataset creation."""
-        from controllers.service_api.dataset.dataset import DatasetListApi
-
-        mock_current_user.__class__ = Account
-        mock_dataset_svc.create_empty_dataset.return_value = Mock()
-        mock_marshal.return_value = {"id": "ds-1", "name": "New Dataset"}
-
-        with app.test_request_context(
-            "/datasets",
-            method="POST",
-            json={"name": "New Dataset"},
-        ):
-            api = DatasetListApi()
-            response, status = _unwrap(api.post)(api, tenant_id=mock_tenant.id)
-
-        assert status == 200
-        mock_dataset_svc.create_empty_dataset.assert_called_once()
-
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_create_dataset_duplicate_name(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        app,
-        mock_tenant,
-    ):
-        """Test DatasetNameDuplicateError when name already exists."""
-        from controllers.service_api.dataset.dataset import DatasetListApi
-
-        mock_current_user.__class__ = Account
-        mock_dataset_svc.create_empty_dataset.side_effect = services.errors.dataset.DatasetNameDuplicateError()
-
-        with app.test_request_context(
-            "/datasets",
-            method="POST",
-            json={"name": "Existing Dataset"},
-        ):
-            api = DatasetListApi()
-            with pytest.raises(DatasetNameDuplicateError):
-                _unwrap(api.post)(api, tenant_id=mock_tenant.id)
-
-
-class TestDatasetApiGet:
-    """Test suite for DatasetApi.get() endpoint.
-
-    ``get`` has no billing decorators but calls ``DatasetService``,
-    ``create_plugin_provider_manager``, ``marshal``, and ``current_user``.
-    """
-
-    @patch("controllers.service_api.dataset.dataset.DatasetPermissionService")
-    @patch("controllers.service_api.dataset.dataset.marshal")
-    @patch("controllers.service_api.dataset.dataset.create_plugin_provider_manager")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_get_dataset_success(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        mock_provider_mgr,
-        mock_marshal,
-        mock_perm_svc,
-        app,
-        mock_dataset,
-    ):
-        """Test successful dataset retrieval."""
-        from controllers.service_api.dataset.dataset import DatasetApi
-
-        mock_dataset_svc.get_dataset.return_value = mock_dataset
-        mock_dataset_svc.check_dataset_permission.return_value = None
-        mock_current_user.__class__ = Account
-        mock_current_user.current_tenant_id = mock_dataset.tenant_id
-
-        mock_configs = Mock()
-        mock_configs.get_models.return_value = []
-        mock_provider_mgr.return_value.get_configurations.return_value = mock_configs
-
-        mock_marshal.return_value = {
-            "indexing_technique": "economy",
-            "embedding_model_provider": None,
-            "permission": "only_me",
-        }
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}",
-            method="GET",
-        ):
-            api = DatasetApi()
-            response, status = api.get(_=mock_dataset.tenant_id, dataset_id=mock_dataset.id)
-
-        assert status == 200
-        assert response["embedding_available"] is True
-
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_get_dataset_not_found(self, mock_dataset_svc, app, mock_dataset):
-        """Test 404 when dataset not found."""
-        from controllers.service_api.dataset.dataset import DatasetApi
-
-        mock_dataset_svc.get_dataset.return_value = None
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}",
-            method="GET",
-        ):
-            api = DatasetApi()
-            with pytest.raises(NotFound):
-                api.get(_=mock_dataset.tenant_id, dataset_id=mock_dataset.id)
-
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_get_dataset_no_permission(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        app,
-        mock_dataset,
-    ):
-        """Test 403 when user has no permission."""
-        from controllers.service_api.dataset.dataset import DatasetApi
-
-        mock_dataset_svc.get_dataset.return_value = mock_dataset
-        mock_dataset_svc.check_dataset_permission.side_effect = services.errors.account.NoPermissionError()
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}",
-            method="GET",
-        ):
-            api = DatasetApi()
-            with pytest.raises(Forbidden):
-                api.get(_=mock_dataset.tenant_id, dataset_id=mock_dataset.id)
-
-
-class TestDatasetApiDelete:
-    """Test suite for DatasetApi.delete() endpoint.
-
-    ``delete`` is wrapped by ``@cloud_edition_billing_rate_limit_check``.
-    """
-
-    @patch("controllers.service_api.dataset.dataset.DatasetPermissionService")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_delete_dataset_success(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        mock_perm_svc,
-        app,
-        mock_dataset,
-    ):
-        """Test successful dataset deletion."""
-        from controllers.service_api.dataset.dataset import DatasetApi
-
-        mock_dataset_svc.delete_dataset.return_value = True
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}",
-            method="DELETE",
-        ):
-            api = DatasetApi()
-            result = _unwrap(api.delete)(api, _=mock_dataset.tenant_id, dataset_id=mock_dataset.id)
-
-        assert result == ("", 204)
-
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_delete_dataset_not_found(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        app,
-        mock_dataset,
-    ):
-        """Test 404 when dataset not found for deletion."""
-        from controllers.service_api.dataset.dataset import DatasetApi
-
-        mock_dataset_svc.delete_dataset.return_value = False
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}",
-            method="DELETE",
-        ):
-            api = DatasetApi()
-            with pytest.raises(NotFound):
-                _unwrap(api.delete)(api, _=mock_dataset.tenant_id, dataset_id=mock_dataset.id)
-
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_delete_dataset_in_use(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        app,
-        mock_dataset,
-    ):
-        """Test DatasetInUseError when dataset is in use."""
-        from controllers.service_api.dataset.dataset import DatasetApi
-
-        mock_dataset_svc.delete_dataset.side_effect = services.errors.dataset.DatasetInUseError()
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}",
-            method="DELETE",
-        ):
-            api = DatasetApi()
-            with pytest.raises(DatasetInUseError):
-                _unwrap(api.delete)(api, _=mock_dataset.tenant_id, dataset_id=mock_dataset.id)
-
-
-class TestDocumentStatusApiPatch:
-    """Test suite for DocumentStatusApi.patch() endpoint.
-
-    ``patch`` has no billing decorators but calls ``DatasetService``,
-    ``DocumentService``, and ``current_user``.
-    """
-
-    @patch("controllers.service_api.dataset.dataset.DocumentService")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_batch_update_status_success(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        mock_doc_svc,
-        app,
-        mock_tenant,
-        mock_dataset,
-    ):
-        """Test successful batch document status update."""
-        from controllers.service_api.dataset.dataset import DocumentStatusApi
-
-        mock_current_user.__class__ = Account
-        mock_dataset_svc.get_dataset.return_value = mock_dataset
-        mock_dataset_svc.check_dataset_permission.return_value = None
-        mock_dataset_svc.check_dataset_model_setting.return_value = None
-        mock_doc_svc.batch_update_document_status.return_value = None
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}/documents/status/enable",
-            method="PATCH",
-            json={"document_ids": ["doc-1", "doc-2"]},
-        ):
-            api = DocumentStatusApi()
-            response, status = api.patch(
-                tenant_id=mock_tenant.id,
-                dataset_id=mock_dataset.id,
-                action="enable",
-            )
-
-        assert status == 200
-        assert response["result"] == "success"
-
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_batch_update_status_dataset_not_found(
-        self,
-        mock_dataset_svc,
-        app,
-        mock_tenant,
-        mock_dataset,
-    ):
-        """Test 404 when dataset not found."""
-        from controllers.service_api.dataset.dataset import DocumentStatusApi
-
-        mock_dataset_svc.get_dataset.return_value = None
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}/documents/status/enable",
-            method="PATCH",
-            json={"document_ids": ["doc-1"]},
-        ):
-            api = DocumentStatusApi()
-            with pytest.raises(NotFound):
-                api.patch(
-                    tenant_id=mock_tenant.id,
-                    dataset_id=mock_dataset.id,
-                    action="enable",
-                )
-
-    @patch("controllers.service_api.dataset.dataset.DocumentService")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_batch_update_status_indexing_error(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        mock_doc_svc,
-        app,
-        mock_tenant,
-        mock_dataset,
-    ):
-        """Test InvalidActionError when document is indexing."""
-        from controllers.service_api.dataset.dataset import DocumentStatusApi
-
-        mock_current_user.__class__ = Account
-        mock_dataset_svc.get_dataset.return_value = mock_dataset
-        mock_dataset_svc.check_dataset_permission.return_value = None
-        mock_dataset_svc.check_dataset_model_setting.return_value = None
-        mock_doc_svc.batch_update_document_status.side_effect = services.errors.document.DocumentIndexingError()
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}/documents/status/enable",
-            method="PATCH",
-            json={"document_ids": ["doc-1"]},
-        ):
-            api = DocumentStatusApi()
-            with pytest.raises(InvalidActionError):
-                api.patch(
-                    tenant_id=mock_tenant.id,
-                    dataset_id=mock_dataset.id,
-                    action="enable",
-                )
-
-    @patch("controllers.service_api.dataset.dataset.DocumentService")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    @patch("controllers.service_api.dataset.dataset.DatasetService")
-    def test_batch_update_status_value_error(
-        self,
-        mock_dataset_svc,
-        mock_current_user,
-        mock_doc_svc,
-        app,
-        mock_tenant,
-        mock_dataset,
-    ):
-        """Test InvalidActionError when ValueError raised."""
-        from controllers.service_api.dataset.dataset import DocumentStatusApi
-
-        mock_current_user.__class__ = Account
-        mock_dataset_svc.get_dataset.return_value = mock_dataset
-        mock_dataset_svc.check_dataset_permission.return_value = None
-        mock_dataset_svc.check_dataset_model_setting.return_value = None
-        mock_doc_svc.batch_update_document_status.side_effect = ValueError("Invalid action")
-
-        with app.test_request_context(
-            f"/datasets/{mock_dataset.id}/documents/status/enable",
-            method="PATCH",
-            json={"document_ids": ["doc-1"]},
-        ):
-            api = DocumentStatusApi()
-            with pytest.raises(InvalidActionError):
-                api.patch(
-                    tenant_id=mock_tenant.id,
-                    dataset_id=mock_dataset.id,
-                    action="enable",
-                )
-
-
-class TestDatasetTagsApiGet:
-    """Test suite for DatasetTagsApi.get() endpoint."""
-
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    def test_list_tags_success(
-        self,
-        mock_current_user,
-        mock_tag_svc,
-        app,
-    ):
-        """Test successful tag list retrieval."""
-        from controllers.service_api.dataset.dataset import DatasetTagsApi
-
-        mock_current_user.__class__ = Account
-        mock_current_user.current_tenant_id = "tenant-1"
-        mock_tag = SimpleNamespace(id="tag-1", name="Test Tag", type="knowledge", binding_count="0")
-        mock_tag_svc.get_tags.return_value = [mock_tag]
-
-        with app.test_request_context("/datasets/tags", method="GET"):
-            api = DatasetTagsApi()
-            response, status = api.get(_=None)
-
-        assert status == 200
-        assert len(response) == 1
-
-
-class TestDatasetTagsApiPost:
-    """Test suite for DatasetTagsApi.post() endpoint."""
-
-    # BUG: dataset.py L512 passes ``binding_count=0`` (int) to
-    # ``DataSetTag.model_validate()``, but ``DataSetTag.binding_count``
-    # is typed ``str | None`` (see fields/tag_fields.py L20).
-    # This causes a Pydantic ValidationError at runtime.
-    @pytest.mark.skip(reason="Production bug: DataSetTag.binding_count is str|None but dataset.py passes int 0")
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    def test_create_tag_success(
-        self,
-        mock_current_user,
-        mock_tag_svc,
-        app,
-    ):
-        """Test successful tag creation."""
-        from controllers.service_api.dataset.dataset import DatasetTagsApi
-
-        mock_current_user.__class__ = Account
-        mock_current_user.has_edit_permission = True
-        mock_current_user.is_dataset_editor = True
-        mock_tag = SimpleNamespace(id="tag-new", name="New Tag", type="knowledge")
-        mock_tag_svc.save_tags.return_value = mock_tag
-
-        with app.test_request_context(
-            "/datasets/tags",
-            method="POST",
-            json={"name": "New Tag"},
-        ):
-            api = DatasetTagsApi()
-            response, status = api.post(_=None)
-
-        assert status == 200
-        assert response["name"] == "New Tag"
-        mock_tag_svc.save_tags.assert_called_once()
-
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    def test_create_tag_forbidden(self, mock_current_user, app):
-        """Test 403 when user lacks edit permission."""
-        from controllers.service_api.dataset.dataset import DatasetTagsApi
-
-        mock_current_user.__class__ = Account
-        mock_current_user.has_edit_permission = False
-        mock_current_user.is_dataset_editor = False
-
-        with app.test_request_context(
-            "/datasets/tags",
-            method="POST",
-            json={"name": "New Tag"},
-        ):
-            api = DatasetTagsApi()
-            with pytest.raises(Forbidden):
-                api.post(_=None)
-
-
-class TestDatasetTagBindingApiPost:
-    """Test suite for DatasetTagBindingApi.post() endpoint."""
-
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    def test_bind_tags_success(
-        self,
-        mock_current_user,
-        mock_tag_svc,
-        app,
-    ):
-        """Test successful tag binding."""
-        from controllers.service_api.dataset.dataset import DatasetTagBindingApi
-
-        mock_current_user.__class__ = Account
-        mock_current_user.has_edit_permission = True
-        mock_current_user.is_dataset_editor = True
-        mock_tag_svc.save_tag_binding.return_value = None
-
-        with app.test_request_context(
-            "/datasets/tags/binding",
-            method="POST",
-            json={"tag_ids": ["tag-1"], "target_id": "ds-1"},
-        ):
-            api = DatasetTagBindingApi()
-            result = api.post(_=None)
-
-        assert result == ("", 204)
-
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    def test_bind_tags_forbidden(self, mock_current_user, app):
-        """Test 403 when user lacks edit permission."""
-        from controllers.service_api.dataset.dataset import DatasetTagBindingApi
-
-        mock_current_user.__class__ = Account
-        mock_current_user.has_edit_permission = False
-        mock_current_user.is_dataset_editor = False
-
-        with app.test_request_context(
-            "/datasets/tags/binding",
-            method="POST",
-            json={"tag_ids": ["tag-1"], "target_id": "ds-1"},
-        ):
-            api = DatasetTagBindingApi()
-            with pytest.raises(Forbidden):
-                api.post(_=None)
-
-
-class TestDatasetTagUnbindingApiPost:
-    """Test suite for DatasetTagUnbindingApi.post() endpoint."""
-
-    @patch("controllers.service_api.dataset.dataset.TagService")
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    def test_unbind_tag_success(
-        self,
-        mock_current_user,
-        mock_tag_svc,
-        app,
-    ):
-        """Test successful tag unbinding."""
-        from controllers.service_api.dataset.dataset import DatasetTagUnbindingApi
-
-        mock_current_user.__class__ = Account
-        mock_current_user.has_edit_permission = True
-        mock_current_user.is_dataset_editor = True
-        mock_tag_svc.delete_tag_binding.return_value = None
-
-        with app.test_request_context(
-            "/datasets/tags/unbinding",
-            method="POST",
-            json={"tag_id": "tag-1", "target_id": "ds-1"},
-        ):
-            api = DatasetTagUnbindingApi()
-            result = api.post(_=None)
-
-        assert result == ("", 204)
-
-    @patch("controllers.service_api.dataset.dataset.current_user")
-    def test_unbind_tag_forbidden(self, mock_current_user, app):
-        """Test 403 when user lacks edit permission."""
-        from controllers.service_api.dataset.dataset import DatasetTagUnbindingApi
-
-        mock_current_user.__class__ = Account
-        mock_current_user.has_edit_permission = False
-        mock_current_user.is_dataset_editor = False
-
-        with app.test_request_context(
-            "/datasets/tags/unbinding",
-            method="POST",
-            json={"tag_id": "tag-1", "target_id": "ds-1"},
-        ):
-            api = DatasetTagUnbindingApi()
-            with pytest.raises(Forbidden):
-                api.post(_=None)
